requirements.txt
aniso8601==8.1.0
appdirs==1.4.4
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
flask-restx==0.2.0
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
jsonschema==3.2.0
MarkupSafe==1.1.1
pymongo==3.11.2
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.4.0
Werkzeug==1.0.1
zipp==3.4.0
